01_Data Extraction/requirements.txt
python-dotenv
serpapi
pandas